+++ jsonargparse/_deprecated.py
@@ -14,7 +14,7 @@ from typing import Any, Callable, Dict, Optional, Set
 from ._common import Action, null_logger
 from ._common import LoggerProperty as InternalLoggerProperty
 from ._namespace import Namespace
-from ._type_checking import ArgumentParser, ruyamlCommentedMap
+from ._type_checking import ArgumentParser, ruamelCommentedMap
 
 __all__ = [
     "ActionEnum",
@@ -711,34 +711,34 @@ class HelpFormatterDeprecations:
         return self._yaml_formatter.add_yaml_comments(cfg)
 
     @deprecated("The set_yaml_start_comment method is deprecated and will be removed in v5.0.0.")
-    def set_yaml_start_comment(self, text: str, cfg: ruyamlCommentedMap):
-        """Sets the start comment to a ruyaml object.
+    def set_yaml_start_comment(self, text: str, cfg: ruamelCommentedMap):
+        """Sets the start comment to a ruamel.yaml object.
 
         Args:
             text: The content to use for the comment.
-            cfg: The ruyaml object.
+            cfg: The ruamel.yaml object.
         """
         self._yaml_formatter.set_yaml_start_comment(text, cfg)
 
     @deprecated("The set_yaml_group_comment method is deprecated and will be removed in v5.0.0.")
-    def set_yaml_group_comment(self, text: str, cfg: ruyamlCommentedMap, key: str, depth: int):
-        """Sets the comment for a group to a ruyaml object.
+    def set_yaml_group_comment(self, text: str, cfg: ruamelCommentedMap, key: str, depth: int):
+        """Sets the comment for a group to a ruamel.yaml object.
 
         Args:
             text: The content to use for the comment.
-            cfg: The parent ruyaml object.
+            cfg: The parent ruamel.yaml object.
             key: The key of the group.
             depth: The nested level of the group.
         """
         self._yaml_formatter.set_yaml_group_comment(text, cfg, key, depth)
 
     @deprecated("The set_yaml_argument_comment method is deprecated and will be removed in v5.0.0.")
-    def set_yaml_argument_comment(self, text: str, cfg: ruyamlCommentedMap, key: str, depth: int):
-        """Sets the comment for an argument to a ruyaml object.
+    def set_yaml_argument_comment(self, text: str, cfg: ruamelCommentedMap, key: str, depth: int):
+        """Sets the comment for an argument to a ruamel.yaml object.
 
         Args:
             text: The content to use for the comment.
-            cfg: The parent ruyaml object.
+            cfg: The parent ruamel.yaml object.
             key: The key of the argument.
             depth: The nested level of the argument.
         """