+++ jsonargparse/_type_checking.py
@@ -5,12 +5,12 @@ __all__ = [
     "ActionsContainer",
     "ArgumentParser",
     "docstring_parser",
-    "ruyamlCommentedMap",
+    "ruamelCommentedMap",
 ]
 
 if TYPE_CHECKING:  # pragma: no cover
     import docstring_parser
-    from ruyaml.comments import CommentedMap as ruyamlCommentedMap
+    from ruamel.yaml.comments import CommentedMap as ruamelCommentedMap
 
     from ._core import ActionsContainer, ArgumentGroup, ArgumentParser
 else: